#UDV2-3307 API - Place order and send info to SQS

+++ src/DXP.SmartConnect.Ecom.Core/Entities/OrderSummary.cs
@@ -7,7 +7,7 @@ namespace DXP.SmartConnect.Ecom.Core.Entities
         public int ItemCount { set; get; }
         public string TaxTotal { set; get; }
         public string ServiceFee { set; get; }
-        public int PointsEarned { set; get; }
+        public decimal PointsEarned { set; get; }
         public string CustomerCredit { set; get; }
         public string Total { set; get; }
         public string OrderValue { set; get; }
